// src/app/maps/interfaces/ipTracker.ts
export interface  IPTracker{
    status:      string;
    country:     string;
    countryCode: string;
    region:      string;
    regionName:  string;
    city:        string;
    zip:         string;
    lat:         number;
    lon:         number;
    timezone:    string;
    isp:         string;
    org:         string;
    as:          string;
    query:       string;
}